shared/requirements.txt
Production dependencies for the Shared Clients & Models module

Versions are pinned for stable deployment.

Core GCP services

google-cloud-firestore==2.17.0
google-cloud-pubsub==2.19.0

Data validation and schema

pydantic==2.8.2